pws cloud foundry deployment

+++ 06_bookABatterySERVICE4EUREKA/src/main/java/it/luigibennardis/microservice/repositories/StazioniNamedQueryRepository.java
@@ -9,8 +9,8 @@ import org.springframework.data.repository.query.Param;
 public interface StazioniNamedQueryRepository extends Repository<StationAddress, Long> {
 		 
 	@Query 
-	public List<StationAddress> findNearest(@Param(value = "latitude") long latitude, 
-		@Param(value = "longitude") long longitude, @Param(value = "distance") Double distance );
+	public List<StationAddress> findNearest(@Param(value = "latitude") Double latitude, 
+		@Param(value = "longitude") Double longitude, @Param(value = "distance") Double distance );
 	
 	 
 	}
